greenflow/requirements.txt
pathway
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.20.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.30
tiktoken>=0.6.0
chromadb>=0.4.24
sentence-transformers>=2.7.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
aiosqlite>=0.20.0
redis>=5.0.3
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
httpx>=0.27.0
aiohttp>=3.9.4
websockets>=12.0
aiofiles>=23.2.1
pandas>=2.2.1
numpy>=1.26.4
loguru>=0.7.2
prometheus-fastapi-instrumentator>=6.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.1.0
pytest-asyncio>=0.23.6
black>=24.3.0
isort>=5.13.2
flake8>=7.0.0
streamlit>=1.30.0
plotly>=5.18.0
streamlit-autorefresh>=0.0.1
psutil>=5.9.8

# Google Gemini
langchain-google-genai>=0.0.11
google-generativeai>=0.5.0
